crimson: do not use small_vector
if `std::is_nothrow_move_constructible<small_vector<peer_shard_t, 2>>`
also depends on if the allocator throws when moving from a small_vector
to a new one, as we need to allocate memory for the destination
small_vector.

so, in this change, the optimization for small_vector is dropped,
because the default allocator used by small_vector throws.

Signed-off-by: Kefu Chai <kchai@redhat.com>