msg/async/dpdk: destroy fd in do_request
when testing with dpdk,there is an error log "Event (0xaaaaff889288
nevent = 5000 time_id = 21540). delete_time_event id = 21471 not found",
raised in EventCenter::delete_time_event. In EventCenter::process_time_events,
the fd in event_map is deleted, but fd is not destroyed, and then fd will
not found in delete_time_event. Destroy fd in do_request to ensure
that fd is zero, and then delete_time_event (fd) will not be called.

Signed-off-by: Chunsong Feng <fengchunsong@huawei.com>
Signed-off-by: luo rixin <luorixin@huawei.com>