rbd-mirror: register pool replayer only if init succeeded
So if the init failed it could be retried after refresh
interval. The init error may be temporary, e.g. because the
remote pool is not created yet.

This also fixes a potential crash after a failed pool replayer
was added and then 'rbd mirror status' asok command was executed,
trying to access the pool replayer's uninitialized objects.

Signed-off-by: Mykola Golub <mgolub@suse.com>