mgr/dashboard: Introduce layout components
This PR will simplify the code of the app component in that way that it introduces layout components. Depending on the route a specific layout is choosen. Currently there are two layouts:

- Workbench
- Blank

The blank layout, which does not show any navigation controls, is used for the error, login and logout pages. The workbench layout shows the navigation controls and is mainly used for all pages that are used to configure Ceph.

The 403 and 404 pages have a 'Back' button that will redirect to /login. The Angular router will redirect to /dashboard if logged in.

Fixes: https://tracker.ceph.com/issues/43565

Signed-off-by: Volker Theile <vtheile@suse.com>