qa: save MDS epoch barrier
Rather than verify the latest OSDMap is the same as the rank's
osdmap_epoch_barrier, just use the rank's version. The OSDMap may change
out-of-band with the test startup and thus the epoch's would diverge.
The file system and rank is fresh for each test so there's no reason to
care if the MDS barrier is one epoch behind the latest.

Fixes: https://tracker.ceph.com/issues/43554
Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>