mon/ConfigMonitor: always prefix global config keys with global/
This avoids some ambiguity.  For instance, we cannot trivially tell if

 config/mgr/module/option

is an option called module/option that we don't understand that's in
the mgr section, or a mgr/module/option option in global.  In any case,
it's simpler to have a universal prefix.

Also, this aligns the storage schema with what assimilate-conf does (it
adds global/ for the global section).

Fixes: https://tracker.ceph.com/issues/43296
Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit 9da5cb75c6b892b33af9a4878a9262d687d785bf)