src/: use transparent comparator
use transparent comparator whenever appropriate. transparent comparator
allows us to do heterogeneous lookup. which avoid the overhead to create
a temporary object just for compare it. for instance, we use std::string
for key, without transparent comparator, we will have to create a
std::string instance even if we have a string_view already. after this
change, we can just use a string_view as key when looking up it in the
map<string,val_type>.

Signed-off-by: Kefu Chai <kchai@redhat.com>