doc: change "monitor" to "node in bootstrap.rst
This commit changes prompts that read "monitor" to prompts
that read "node" in commands issued before nodes have become
monitors. (I had thought that this was already done, but
I suspect that in fact I had not understood Sage's email to
me on this subject until sometime after I had read it, and
having understood it just assumed that it was done.) Anyway,
the commands that acquire and run ceph-daemon have now been
situated after prompts that read "node", in order to avoid
making it look chicken-or-egg-like in our documentation as
though a monitor can be prior to itself.

Signed-off-by: Zac Dover <zac.dover@gmail.com>