mgr/cephadm: fix detection of just-created OSDs
Don't match against the literal device names that were passed, as this
will fail if there is (1) an LV name or (2) non-canonical name like
/dev/disk/...

Instead, match any OSD we see that didn't exist before we ran c-v prepare.
This might include other OSDs created by a racing c-v incantation, but (1)
we shouldn't have any of those, and (2) even if they were, the OSD is new
and should be added.

Signed-off-by: Sage Weil <sage@redhat.com>