mon: elector: use unique_ptr on [stable|leader]peer_tracker, fix a leak
We were conditionally taking over a ConnectionTracker which had been
allocated on the heap when recieving a propose, but doing nothing with
the input if we didn't wan it. Callers didn't (couldn't know to) clean these up.
Instead, just copy them into a new ConnectionTracker on use and reset
the existing one.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>