ceph-volume: fix lvm list
fe497bd3590fbcccbc318543bd0073a135c03ba7 introduced a regression in `lvm
list`.

When passing a vg/lv path for generating a single report, it fails
because the filter used in the `lvs` command isn't right. It uses the lv
name instead of the vg name because `os.path.basename(device)` is used
while it should be `os.path.dirname(device)`

Fixes: https://tracker.ceph.com/issues/43969

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>