mimic: common/config: parse --log-early but ignore it
Ideally we woudl backport 933d5084cb66f299a7bf60f0a2a6382c0bd3cb2f, but
it does not apply cleanly to mimic.  Among other things, we modifies
behavior that changed in cd6a5b9c40779956629803f222c365bdb291a169 and
the related commits, which is non-trivial and not worth backporting.

We want this in mimic anyway because ceph_manager.py is now passing it
for all of the raw_cluster_cmds, and it would be even more awkward to
make that behavior conditional on version.

Fixes: https://tracker.ceph.com/issues/43946
Signed-off-by: Sage Weil <sage@redhat.com>