msg/Policy: limit unregistered anon connections to mon
When we were fixing 'tell mon' we changed the messenger to allow multiple
lossy connections to the same server.  This was needed for the mon, and
assumed to be harmless for everyone else.  However, for the OSD, it can
lead to out-of-order requests, as observed in this bug:
	https://tracker.ceph.com/issues/42328

Fix by reverting this behavior change except for the mon:

- Introduce a stateless_anon_server policy that sets
  register_lossy_clients = false

- Effectively revert the behavior change in 11534a27068843cbadf03d8fd212922a491d7996
  so that lossy clients *are* registered for stateless_server policy.

- Use the state_anon_server policy for the mon, which was the one place
  where we needed unregistered connections (for 'tell mon.x' to work).

Fixes: https://tracker.ceph.com/issues/42328
Signed-off-by: Sage Weil <sage@redhat.com>