rbd-mirror: don't expect image map is always initialized
in a namespace replayer when handling "instances added/removed"
event. It is possible an event comes when a pool replayer has
already created and initialized a new namespace replayer but has
not acquired leader for it (image_map is not initialized yet).

Also, make sure `m_image_map` is set only after its
initialization is complete.

Fixes: https://tracker.ceph.com/issues/44161
Signed-off-by: Mykola Golub <mgolub@suse.com>