mon: stash newer map on bootstrap when addr doesn't match
If we have to respawn because a newer monmap comes along where our addr or
rank changes, we need to use that on restart in order to make progress.
Stash the newer map in a temporary location and use it when we restart.

Don't bother cleaning up.  Having this map here is harmless, since we
only use it if it is newer than what is in paxos.

Fixes: https://tracker.ceph.com/issues/44076
Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit 71befa298658925555841a2a98802210954358ce)