mount.ceph: add "fs=<fs_name>" mount options support
"fs" is one alias for "mds_namespace=" and it will be cleaner and
be more user-friendly to use. The "fs" will be translated to
"mds_namespace" before sending it to kernel space.

And the "mds_namespace" will be deprecated to use any more.
Update the documents at the same time.

Fixes: https://tracker.ceph.com/issues/44214
Signed-off-by: Xiubo Li <xiubli@redhat.com>