DNM mds: share custom timeouts with clients by hacking the MDSMap
So, clients don't recognize custom timeouts except when set themselves. This is
fine if the custom timeout is longer than normal, but ultra-bad if it's shorter.

For testing purposes, hack around this by shoving custom timeouts into the MDSMap,
and then encoding the custom value into the global session_timeout value clients
actually look at when a map is destined for them.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>