osd/PeeringState.h: ignore RemoteBackfillReserved in WaitLocalBackfillReserved
It is possible to dequeue an outstanding RemoteBackfillReserved, though we may have
already released reservations for that backfill target. Currently, if this happens
while we are in WaitLocalBackfillReserved, it can lead to a crash on the primary.
Prevent this by treating this condition as a no-op.

The longer term fix is to add a RELEASE_ACK mechanism, which prevents the primary
from scheduling a backfill retry until all the RELEASE_ACKs have been received.

Fixes: https://tracker.ceph.com/issues/44248
Signed-off-by: Neha Ojha <nojha@redhat.com>
(cherry picked from commit 61152ac29654a819eb961e5fc2d7f63b706e15c9)

Conflicts:
	src/osd/PeeringState.h
	  - The RemoteBackfillReserved still in src/osd/PG.h in nautilus.
            Manually backport this fix to src/osd/PG.h