qa/tasks/ceph_manager: increase CLI command timeout
There is a problem with mimic releases where pg_creates can stall for a
long time doing the build_pg_history while holding osd_lock.  That lock
is also used by the tell command processing queue, which means that
commands like 'flush_pg_stats' can block for long periods...and time out.

This is currently happening with mimic->nautilus upgrades.  Note that
the problem is mostly fixed in nautilus and totally fixed in octopus, so
this is just a matter of tolerating slow behavior in old releases for the
purposes of the upgrade tests.

Work around this by increasing the timeout from 120s -> 900s.

Fixes: https://tracker.ceph.com/issues/43914
Signed-off-by: Sage Weil <sage@redhat.com>