osd/PeeringState.h: ignore MLogRec in Peering/GetInfo
It is possible to receive a response for an old GetLog request, after
we've transitioned out of Started/Primary/Peering/GetLog due to a map change.
Such cases are already handled in Nautilus and beyond, due to
168e20ab8b8da3a5aed41b73f9627d10971be67b, old_peering_evt() will detect
such old responses by comparing last_peering_set with reply_epoch and
query_epoch. In Mimic, query_epoch is not reliable and we may incorrectly
start processing the MLogRec and crash the osd.

Fixes: https://tracker.ceph.com/issues/44022
Signed-off-by: Neha Ojha <nojha@redhat.com>