cephadm: --pid=host
Ceph currently has a 'nonce' portion of the entity_addr_t that should
uniquely identify the daemon on a host.  We currently use the pid.  When
daemons are run via cephadm, the pid is always 1, so these are no longer
unique, and things break.

Specifically, with ac7b21bf3fd84f52454f511da67eb7a609e15075 and related
commits, we can't handle multiple incoming connections that do not have
unique addresses.

For now, we can just work around with by using the host pid namespace.

This is a work-around for https://tracker.ceph.com/issues/44358

Signed-off-by: Sage Weil <sage@redhat.com>