mgr/dashboard: refactor vertical menu
This change moves navigation menu structure from the HTML template into
a tree-like data object.

As a result of this, code needed has been reduced and is easier to
extend and maintain. Additionally, OnPush change detection has been
enabled, which reduces overhead in this component by ~300%.

Most functionality is kept except for the following:
- Monitoring menu was hidden when both Prometheus and AlertManager were
not configured. Now it shows the classical banner indicating how to
configure both things (aligned to other components: as RGW, NFS, etc).
- Block menu turned red when there was some error in a block menu entry.

After discussion, it seems that none of these behaviours were
imperative.

Additional changes:
- New BadgeComponent, which allows for displaying a 'badge' element
with customizable properties (class, style, hidden, label) based on the
output of an Observable.
- New MenuComponent, to encapsulate main vertical menu structure.
- New MenuParentItemComponent and MenuItemComponent.
- New FeatureDirective, to allow replacing permissions (and other
    existing and future functionality: like feature toggles or the
    hiding components based on basic/medium/advanced setup).

Fixes: https://tracker.ceph.com/issues/39009
Closes: #27241
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>