src/vcont.py: experimental: add Ceph pkgdeps to a container
This is intended as the beginning as a sort of parallel to vstart.sh.

vcont.py examines the current repo/branch, looks for the 'closest'
parent sha1 repo build on shaman, and uses that repo to identify
package dependencies for ceph (excluding the ceph packages themselves)
and either lists or installs them.  This allows a make install/strip in
that container to drop in development Ceph binaries that will be able
to use the package dependencies, and thus a working Ceph container from
your development source tree.  Currently it doesn't attempt to install
any ceph-container scripts, so the container won't drop in to replace
ceph-container-built images as yet.

Requirements: CentOS 8 container with all the usual build tools
plus python3-requests if not yet installed.

Usage: create a centos8 container, and enter it with your Ceph source
tree mounted (-v hostpath:containerpath).  You may also need to
allow more open files when starting the container (with podman, for
instance, that's --ulimit nofile=4096:4096).  When building Ceph, use
"-DCMAKE_INSTALL_PREFIX=/usr" to cause the right install behavior.

An earlier version of this script was based on yum Python libraries,
because yum didn't have enough functionality; dnf appears to, so this
version simply calls out to dnf, so doesn't absolutely need to be
written in Python.

Signed-off-by: Dan Mick <dan.mick@redhat.com>