mon: switch removed_ranks monmap representation to fix issues
I misunderstood this initially, but we can call remove() on the same rank
multiple times (eg, MonmapMonitor::prepare_join()), and we can't
mark down the same rank multiple times and mean different monitors
because we're still referencing them by their original rank within
a single pending_map.

So just store the removed_ranks as a set. This fixes issues in
the ConnectionTracker, which is the only data structure actually
consuming these so far.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>