mgr/telemetry: force --license when sending while opted-out
Users can manually send telemetry data with 'ceph telemetry send', even
in case they did not opt-in and agree to the license. We ask to
explicitly add '--license' to 'ceph telemetry send' in this case.
This also fixes an issue when opting-out ('ceph telemetry off'), where the
revision was not reset.

Fixes: https://tracker.ceph.com/issues/44442
Signed-off-by: Yaarit Hatuka <yaarit@redhat.com>